requirements.txt
python-pptx
